# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Nikolskiy",
        "description": "Документация API для фронтенда",
        "contact": {
            "email": "balumba.ru@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/level/{id}": {
            "get": {
                "tags": [
                    "other"
                ],
                "summary": "Получение списка сущностей по номеру уровня ТЦ",
                "operationId": "indexLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Номер уровня",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LevelResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/system/rent": {
            "post": {
                "tags": [
                    "system"
                ],
                "summary": "Подача заявки на аренду",
                "operationId": "af391d8e4711b026e4dc78da5228116a",
                "requestBody": {
                    "$ref": "#/components/requestBodies/RentRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscribeResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/restaurant/category/{categorySlug}": {
            "get": {
                "tags": [
                    "restaurant"
                ],
                "summary": "Получение списка кафе/ресторанов в категории",
                "operationId": "categoryItems",
                "parameters": [
                    {
                        "name": "categorySlug",
                        "in": "path",
                        "description": "slug запрашиваемой категории",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию заведения в категории",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер запрашиваемой страницы",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryItemDetailResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurant/list": {
            "get": {
                "tags": [
                    "restaurant"
                ],
                "summary": "Получение списка кафе/ресторанов",
                "operationId": "list",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию заведения",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер запрашиваемой страницы",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryItemDetailResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/service/category/{categorySlug}": {
            "get": {
                "tags": [
                    "service"
                ],
                "summary": "Получение списка сервисов в категории",
                "operationId": "categoryItems",
                "parameters": [
                    {
                        "name": "categorySlug",
                        "in": "path",
                        "description": "slug запрашиваемой категории",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию сервиса/услуги в категории",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер запрашиваемой страницы",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryItemDetailResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/service/list": {
            "get": {
                "tags": [
                    "service"
                ],
                "summary": "Получение списка сервисов",
                "operationId": "list",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию сервиса/услуги",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер запрашиваемой страницы",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryItemDetailResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/shop/category/{categorySlug}": {
            "get": {
                "tags": [
                    "shop"
                ],
                "summary": "Получение списка магазинов в категории",
                "operationId": "categoryItems",
                "parameters": [
                    {
                        "name": "categorySlug",
                        "in": "path",
                        "description": "slug запрашиваемой категории",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию магазина в категории",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер запрашиваемой страницы",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryItemDetailResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RentRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "comment": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SubscribeResponse": {
                "properties": {
                    "result": {
                        "type": "bollean",
                        "default": true
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LevelResponse": {
                "properties": {
                    "type": {
                        "type": "string",
                        "default": "success"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LevelItem"
                        }
                    }
                },
                "type": "object"
            },
            "LevelItem": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "point": {
                        "type": "string"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategorySmall"
                    },
                    "logo": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "default": "shop | restaurant | service"
                    }
                },
                "type": "object"
            },
...
            "CategorySmall": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CategoryItemDetailResponse": {
                "properties": {
                    "type": {
                        "type": "string",
                        "default": "success"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryDetail"
                        }
                    }
                },
                "type": "object"
            },
            "CategoryDetail": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryDetailItem"
                        }
                    },
                    "first_page_url": {
                        "type": "string"
                    },
                    "from": {
                        "type": "integer"
                    },
                    "next_page_url": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "prev_page_url": {
                        "type": "string"
                    },
                    "to": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CategoryDetailItem": {
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "logo": {
                        "type": "string"
                    },
                    "level": {
                        "type": "integer"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategorySmall"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
...
            "RentRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/RentRequest"
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}